feat: handle bookValue as Long in backend, update docker-compose.infra.yml port, and implement customer deletion in table UI

+++ backend/src/main/java/com/app/starter1/dto/ProductoRequest.java
@@ -11,7 +11,7 @@ public class ProductoRequest {
     private String warrantyEndDate;
     private String warrantyStartDate;
     private String supplier;
-    private Integer bookValue;
+    private Long bookValue;
     private String purchaseDate;
     private String amperage;
     private String frequency;
